rename, move and clean up
