+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
-import Layout from "./components/Layout";
+import Layout from "./components/global/Layout";
 import Home from "./pages/home/home";
 import LogIn from "./pages/login/login";
 import Register from "./pages/register/register";
@@ -9,6 +9,7 @@ import Profile from "./pages/profile/profile";
 import RouteNotFound from "./pages/notfound/notfound";
 import Venue from "./pages/specificVenue/specificVenue";
 import NewVenue from "./pages/newVenue/newVenue";
+import UpdateVenue from "./pages/updateVenue/updateVenue";
 import NotLoggedIn from './pages/error';
 
 
@@ -26,6 +27,7 @@ function App() {
         <Route path="/register" element={<Register />} />
         <Route path="/venues" element={<Venues />} />
         <Route path="/newVenue" element={user ? <NewVenue /> : <Navigate replace to={"/error"} />} />
+        <Route path="/updateVenue/:id" element={user ? <UpdateVenue /> : <Navigate replace to={"/error"} />} />
         <Route path="/specific/:id" element={<Venue />} />
         <Route path="/profile" element={user ? <Profile /> : <Navigate replace to={"/error"} />} />
         <Route path="*" element={<RouteNotFound />} />
